refactor(components.charts): move to webUniverseComponents
