+++ client/app/webUniverseComponents/charts/index.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import wucGaugeChartDirective from './gaugeChart/gaugeChart';
+
+const moduleName = 'wucCharts';
+
+angular
+  .module(moduleName, [])
+  .directive('wucGaugeChart', wucGaugeChartDirective);
+
+export default moduleName;
