+++ client/app/webUniverseComponents/index.js
@@ -1,6 +1,7 @@
 import angular from 'angular';
 
 import wucConverter from './converter';
+import wucCharts from './charts';
 import wucCron from './cron';
 import wucCronValidator from './cron-validator';
 import wucFileChange from './fileChange';
@@ -11,6 +12,7 @@ const moduleName = 'webUniverseComponents';
 angular
   .module(moduleName, [
     wucConverter,
+    wucCharts,
     wucCron,
     wucCronValidator,
     wucFileChange,
