+++ package.json
@@ -94,7 +94,7 @@
     "json3": "^3.2.5",
     "jsontoxml": "latest",
     "jsurl": "^0.1.4",
-    "justgage": "Ulflander/justgage#13f8372c4bab2ed576e42199b45ee95dfd14fcf7",
+    "justgage": "^1.2.2",
     "lodash": "^3.10.1",
     "matchdep": "^0.3.0",
     "matchmedia": "^0.1.2",
